Move C DNS wrappers to dns.c
