+++ Sources/files.cmake
@@ -390,6 +390,7 @@ set(PARTOUT_C_SOURCES
 ./MiniFoundation_C/rx.cc
 ./MiniFoundation_C/std.c
 ./PartoutCore_C/common.c
+./PartoutCore_C/dns.c
 ./PartoutCore_C/lib.c
 ./PartoutCore_C/mux.c
 ./PartoutCore_C/network.c
